Fixed script errors in head
There were a couple scripts being called that were causing errors in
Chrome developer tools.

One was calling a local jquery file, but there was already a script
above it that was pulling in jQuery from their website (this way it's
always the most up to date version.

There was also a script calling the bootstrap.min.js file, but it was
also unnecessary since the regular version was already being called
above it.

Anyway, the errors are gone!